# ddl, metrics: speed up the execution of the DDL job