<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #xxx <!-- REMOVE this line if no issue to close -->

Problem Summary:
Speed up the execution of the DDL job, mainly to speed up the table and database creating.

### What is changed and how it works?

Proposal: [xxx](url) <!-- REMOVE this line if not applicable -->

What's Changed:
Relate to issue #14770, we will update the process of "Passively check if a DDL job needs to be processed" to make it actively check.

How it Works:
When we get a new DDL job, we will notify the DDL worker.

### Related changes

- PR to update `pingcap/docs`/`pingcap/docs-cn`:

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Manual test (add detailed scripts or steps below)

Use 2 TiDBs and 3 PDs on a computer.
case0: With the master branch, the owner and the server receiving the DDL request aren't on the same TiDB.
case1: With the master branch, the owner and the server receiving the DDL request are on the same TiDB.
case2: With the current PR, the owner and the server receiving the DDL request aren't on the same TiDB.
| goroutine count  | table count | case0 | case1 | case2 |
|------------------|------------|-------|-------|-------|
|                  1           |          9       |    9.1s  |   1.5s  |   2.1s   |
|                  1           |        99       | 100.8s | 14.8s |  21.6s  |
|                  1           |       999      | 16m39.3s | 2m27.1s |  3m30.7s  |
|                  3           |          9       |    3.4s  |   1.8s  |   2.3s   |
|                  3           |        99       |   33.6s | 14.8s |  17.9s  |
|                  3           |       999      | 5m20.1s | 2m24.1s |  2m59.7s  |


### Release note <!-- bugfixes or new feature need a release note -->

- Speed up the execution of the DDL job, mainly to speed up the table and database creating.