@hi-rustin 2021-03-02
@zimulala 

Try `/merge cancel`